Move conseval sys mem input forcing to the hoisting pass
