+++ include/ttmlir/Dialect/TTNN/Pipelines/TTNNPipelines.h
@@ -372,12 +372,6 @@ struct TTIRToTTNNCommonPipelineOptions
       llvm::cl::desc("Enable hoisting const-eval ops to CPU module."),
       llvm::cl::init(true)};
 
-  // Force const-eval function inputs to system memory.
-  Option<bool> enableConstEvalInputsToSystemMemory{
-      *this, "enable-const-eval-inputs-to-system-memory",
-      llvm::cl::desc("Force const-eval function inputs to system memory."),
-      llvm::cl::init(true)};
-
   Option<bool> enableTrace{*this, "enable-trace",
                            llvm::cl::desc("Enable trace optimization pass."),
                            llvm::cl::init(false)};
