+++ lib/Dialect/TTNN/Transforms/TTNNLayout.cpp
@@ -635,7 +635,7 @@ private:
     }
 
     // If function is marked as const-eval leave inputs as is.
-    // TTNNConstEvalInputsToSystemMemory pass will handle them.
+    // The const-eval hoist transform forces them to system memory.
     if (ttmlir::utils::isConstEvalFunc(owningFunc)) {
       return false;
     }
